Minimum Load Requirements...
This standard provides minimum load requirements for the design of buildings and other structures that are subject to building code requirements. Loads and appropriate load combinations, which have been developed to be used together, are set forth for strength design and allowable stress design. For design strengths and allowable stress limits, design speciﬁ cations for conventional structural materials used in buildings and modiﬁ cations contained in this standard shall be followed.